tweaks & refactor & fix Armature follow_path
